requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
faiss-cpu==1.11.0
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
ollama==0.4.8
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3